Replace the old dispatcher example with the new one
